refactor: remove FamilyMember entity and simplify member-familyroom relation
