--- src/main/java/com/urisik/backend/domain/familyroom/enums/FamilyStatus.java
@@ -1,5 +1,0 @@
-package com.urisik.backend.domain.familyroom.enums;
-
-public enum FamilyStatus {
-    ACTIVE, INACTIVE
-}
