+++ src/main/java/com/urisik/backend/domain/familyroom/repository/FamilyRoomRepository.java
@@ -9,6 +9,4 @@ public interface FamilyRoomRepository extends JpaRepository<FamilyRoom, Long> {
 
     // 맴버가 속한 familyroom 가져오기
     Optional<FamilyRoom> findByMembers_Id(Long memberId);
-
-
 }
