+++ src/main/java/com/urisik/backend/domain/familyroom/repository/InviteRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface InviteRepository extends JpaRepository<Invite, Long> {
+
     Optional<Invite> findByToken(String token);
     boolean existsByToken(String token);
 }
