+++ src/main/java/com/urisik/backend/domain/member/converter/FamilyMemberProfileConverter.java
@@ -35,7 +35,7 @@ public class FamilyMemberProfileConverter {
         return FamilyMemberProfileResponse.Create.builder()
                 .isSuccess(true)
                 .nickname(profile.getNickname())
-                .role(profile.getRole())
+                .role(profile.getFamilyRole())
                 .likedIngredients(profile.getLikedIngredients())
                 .dislikedIngredients(profile.getDislikedIngredients())
                 .allergy(allergies)
@@ -64,7 +64,7 @@ public class FamilyMemberProfileConverter {
         return FamilyMemberProfileResponse.Detail.builder()
                 .isSuccess(true)
                 .nickname(profile.getNickname())
-                .role(profile.getRole())
+                .role(profile.getFamilyRole())
                 .likedIngredients(profile.getLikedIngredients())
                 .dislikedIngredients(profile.getDislikedIngredients())
                 .allergy(allergies)
