+++ src/main/java/com/urisik/backend/domain/member/dto/res/FamilyMemberProfileResponse.java
@@ -1,10 +1,8 @@
 package com.urisik.backend.domain.member.dto.res;
 
 import com.urisik.backend.domain.allergy.enums.Allergen;
-import com.urisik.backend.domain.familyroom.enums.FamilyRole;
+import com.urisik.backend.domain.member.enums.FamilyRole;
 import com.urisik.backend.domain.member.enums.DietPreferenceList;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
